Add main.js and text parser module
The text parser generates an index of sentences categorized by the first
word which they start with.

I could have used 'split' to handle the initial seed text, but I prefer
to step through it one character at a time and parse it that way. This
is lighter on memory usage, and still performs in about linnear
complexity. It also gives some flexibility in parsing different types of
things - for example, I change newlines into spaces as I go. That
would be possible if using 'split', but it would work slightly
differently - it might take an extra step.