Finish main.js and add tweaks to fiction generator
In order to get greater variety of sentences, I added more to the
default words for sentences to start with when the chain is 'broken' so
to speak.

This also removes the generated text 'Test__Pirates' and comments out
all the code except for the generation of the fiction. Since the parsed
seed data is already generated and committed, we don't need to keep
re-generating it every time the script runs.

Now, `npm start` will run the script to generate the novel chapters,
overwriting any previous versions that were present.